pyproject.toml
[tool.poetry]
name = "openclaw-harness"
version = "0.1.0"
description = "Harness Engineering framework for OpenClaw"
authors = ["Flw <fulingwei@gmail.com>"]
readme = "README.md"
packages = [{include = "harness", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0"
click = "^8.0"
jinja2 = "^3.0"
openai = "^1.0"
anthropic = "^0.25"
rich = "^13.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^24.0"
mypy = "^1.0"

[tool.poetry.scripts]
harness = "cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
